Figuring out MainGameScreen architecture.
We have a simple display working under the LML-templated menu. At present, it is not the most exciting thing, as it only draws a green circle.

The idea at present is to define Screens. Only one Screen is active within the MainGameScreen at a time. The MGS executes the Screen's render() method when appropriate. At present, MGS is managing the Screen's Viewport -- but I'm not thinking that's the way forward. Maybe stick with a resize() method.