New ListeningSystem enables a few changes to map-cell/entity management. Also: CityMap now has only 1 Entity associated per map-cell.
ListeningSystem gathers EntityListener events -- i.e., whenever an entity is "added" or "dropped" from a Family. Every tick, that system will notify each Entity that such-and-such has been added or dropped.

This means that we can, e.g., detect map-changes better:
 - when the map changes (say, when a corner's height changes), we add a IsMapCellRearranged component to the Entity associated with that map-cell
 - any relevant Module should register a ListeningSystem, listening for the combination of IsMapCellRearranged and those of its components that are "rearrangement-sensitive"

For example: the terrain-tile module registers a ListeningSystem. When the map is modified, the assigned IsTerrainTile is removed from that cell, forcing the assigned terrain-tile to be recomputed at the next tick.