Revamped how "provided" objects work.
Prior to this, a "provided" object was simply made available as a variable to other downstream Modules.

This makes for a mess, as this results in a whole lot of global variables.

Now -- while each Module has its own global-variables to worry about, at least it doesn't have *other* Modules' globals to worry about.

Instead, a Module may access upstream-Modules via its "modules" Map -- e.g., modules['someID'] -- and that upstream-Module's "provided" objects are available as (unmodifiable) members of that Map.